requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
